Changes requested during code review, including raising an error when a timeout issues happens, rename the button which opens the search widget, as well as a redesigned of ArchiveSearchWidget to handle other search methods.
