+++ src/badger/built_in_plugins/environments/sphere_2d/__init__.py
@@ -1,6 +1,4 @@
 from badger import environment
-from PyQt5.QtNetwork import QNetworkRequest
-from PyQt5.QtCore import QUrl
 
 
 class Environment(environment.Environment):
@@ -35,12 +33,3 @@ class Environment(environment.Environment):
 
     def get_observables(self, observable_names):
         return {k: self._observations[k] for k in observable_names}
-
-    def search(self, keyword):
-        url_string = (
-            f"http://lcls-archapp.slac.stanford.edu/"
-            f"retrieval/bpl/searchForPVsRegex?regex=.*{keyword}.*"
-        )
-        request = QNetworkRequest(QUrl(url_string))
-
-        return request
