+++ src/badger/environment.py
@@ -122,7 +122,7 @@ class Environment(BaseModel, ABC):
     def get_bound(self, variable_name):
         return self.get_bounds([variable_name])[variable_name]
 
-    def search(self, keyword=None, archive_url=None):
+    def search(self, keyword: str, callback: callable):
         return None
 
     ############################################################
